pyproject.toml
[project]
name = "resonance-caster"
version = "0.1.0"
description = "Podcast Platform"
authors = [
    {name = "ji-ho lee",email = "search5@gmail.com"}
]
license = {text = "BSD"}
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "plaster-pastedeploy (>=1.0.1,<2.0.0)",
    "pyramid (>=2.0.2,<3.0.0)",
    "pyramid-jinja2 (>=2.10.1,<3.0.0)",
    "pyramid-debugtoolbar (>=4.12.1,<5.0.0)",
    "waitress (>=3.0.2,<4.0.0)",
    "google-cloud-storage (>=3.1.0,<4.0.0)",
    "google-cloud-firestore (>=2.20.2,<3.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "bcrypt (>=4.3.0,<5.0.0)",
    "pydub (>=0.25.1,<0.26.0)",
    "python-slugify (>=8.0.4,<9.0.0)"
]
packages = [{include = "resonance_caster", from = "src", exclude = ["tests"]}]


# MANIFEST.in에서 변환된 제외 파일 패턴
exclude = [
    "**/__pycache__/**",
    "**/*.py[co]"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.plugins."paste.app_factory"]
main = "resonance_caster:main"

[tool.poetry.group.test.dependencies]
WebTest = "*"
pytest = "*"
pytest-cov = "*"
